src/python/requirements.txt
# Getnet Financial Closure - Python deps for notebooks
# Install on cluster or use in job environment
msal>=1.24.0
requests>=2.28.0
openpyxl>=3.1.0
pandas>=2.0.0
pyyaml>=6.0
